More robust offline mode in service worker
There was weirdness with stale pages and ones that didn't
update when tweaking minor things like css. Thought I fixed it, but
turns out it didn't really work that well. Would still see some unstyled
text in certain circumstances. It's the `staleWhileRevalidate` that does
it.

To fix, I turned to networkFirst, but with a fallback to an offline page
for HTML so it'll continue working right offline. This _seems_ to have
fixed everything.