Vastly improves the purgecss plugin
Jekyll Assets runs async, so on post_write from standard
Jekyll hooks, we can't be sure that our assets are done writing.
We want to run purgecss on our CSS files only after they're
written out entirely. This plugin accomplishes that by hooking
into JekyllAssets:env:after_init and JekyllAssets:env:after_write
I've written an optimization to make sure we don't re-run purgecss
on files that are already done. It makes use of temporary files
that have a hash of the css file contents in it to compute whether
it needs to get changed or not. It'll run for all CSS files
independently since you may create multiple when Jekyll Assets
writes a new version of the file. It prints to console too to
show what's happening.