Rewrites how font levels are calculated (-4KB!)
Font levels had a bunch of redundant info in the outputted CSS - this removes it in favor of a css variable for the base font size that’s calculated based off the device. Also removes redundant css vars never used. Has the added benefit of slightly increased line height on iPad and Desktop