Overhauls about page in favor of vertical content
The about page didn’t flow well and was too chaotic. This simplifies and humanizes it, adding a ton of new content that’s more streamlined. There’s a regular header to match the rest of the pages with a cut down about me, Technologies (was skills), Reading (no separate header for current reads), Online (was social), and Offline (was fun facts)