fix: remove need to import types manually + ignore generated files
Makes sure tsconfig allows for project-level types. Makes linting &
prettier ignore the generated file (which is renamed to .d.ts) so that
it doesn't incorrectly report.

Fixes Contentful to use ids instead of names, which should improve a lot
of the default generated names.

Finally, adds type checking to lint-staged so that we actually do that
on a commit before checking in bad code.